Rooftop PV actuals and forecasts

Returns AEMO's region-level estimate of rooftop PV generation, either actuals or forecasts. Published at 30-minute resolution.

Usage

aemo_rooftop_pv(region, start, end, type = c("actual", "forecast"))

Arguments

region

NEM region code.

start, end

Window.

type

One of "actual" (default) or "forecast".

Value

An aemo_tbl.

Details

The "actual" figure is an AEMO estimate derived from the APVI sampling model and weather data, not metered SCADA output. It is the best available public measure of aggregate rooftop PV generation but is subject to revision.

See also

Examples

# \donttest{
op <- options(aemo.cache_dir = tempdir())
try({
  now <- Sys.time()
  r <- aemo_rooftop_pv("NSW1",
                        start = now - 3600, end = now)
  head(r)
})
#> ℹ Downloading <http://nemweb.com.au/Reports/Current/ROOFTOP_PV/ACTUAL/PUBLIC_RO…
#> ✔ Downloading <http://nemweb.com.au/Reports/Current/ROOFTOP_PV/ACTUAL/PUBLIC_RO…
#> 
#> ℹ Downloading <http://nemweb.com.au/Reports/Current/ROOFTOP_PV/ACTUAL/PUBLIC_RO…
#> ✔ Downloading <http://nemweb.com.au/Reports/Current/ROOFTOP_PV/ACTUAL/PUBLIC_RO…
#> 
#> ℹ Downloading <http://nemweb.com.au/Reports/Current/ROOFTOP_PV/ACTUAL/PUBLIC_RO…
#> ✔ Downloading <http://nemweb.com.au/Reports/Current/ROOFTOP_PV/ACTUAL/PUBLIC_RO…
#> 
#> ℹ Downloading <http://nemweb.com.au/Reports/Current/ROOFTOP_PV/ACTUAL/PUBLIC_RO…
#> ✔ Downloading <http://nemweb.com.au/Reports/Current/ROOFTOP_PV/ACTUAL/PUBLIC_RO…
#> 
#> # aemo_tbl: AEMO rooftop PV actual NSW1
#> # Source:   http://nemweb.com.au
#> # Licence:  AEMO Copyright Permissions Notice
#> # Retrieved: 2026-05-04 19:27 UTC 
#> # Rows: 2  Cols: 7
#> 
#>     interval_datetime regionid power  qi        type         lastchanged
#> 1 2026-05-05 04:30:00     NSW1     0   1 MEASUREMENT 2026/05/05 04:49:05
#> 2 2026-05-05 04:30:00     NSW1     0 0.6   SATELLITE 2026/05/05 04:50:18
#>        settlementdate
#> 1 2026-05-05 04:30:00
#> 2 2026-05-05 04:30:00
options(op)
# }